Move eval-log-viewer.ts to query-evaluation-logging/eval-log-viewer.ts and related tests
