+++ extensions/ql-vscode/src/query-history/query-history-manager.ts
@@ -44,7 +44,7 @@ import { CliVersionConstraint } from "../cli";
 import { HistoryItemLabelProvider } from "./history-item-label-provider";
 import { ResultsView } from "../interface";
 import { WebviewReveal } from "../interface-utils";
-import { EvalLogViewer } from "../eval-log-viewer";
+import { EvalLogViewer } from "../query-evaluation-logging/eval-log-viewer";
 import EvalLogTreeBuilder from "../query-evaluation-logging/eval-log-tree-builder";
 import { EvalLogData, parseViewerData } from "../pure/log-summary-parser";
 import { QueryWithResults } from "../run-queries-shared";
